requirements.txt
flask==1.1.2
keras==2.4.3
tensorflow==2.3.0
pillow==7.1.2
matplotlib==3.2.2
